Fix Exponential growth in Daily simulation
Repeated calls to this procedure cause exponential growth in orders. Including this condition fixes that.

During the data simulation process, there is a process to pick orders. If there is not enough stock to pick a full order, that order is split into an Original and Backordered Order. This could result in a chain of orders, with only the final one having a null value for the BackorderOrderID.